continue..

+++ src/database/MysqlClient.ts
@@ -3,7 +3,7 @@ import { SomeDataBase } from "../types/SomeDataBase";
 import mysql from 'mysql';
 import { Pool, PoolConnection } from "mysql";
 import dotenv from 'dotenv';
-import { AllCategory, ErrorInsertInto, Product } from '../types/Product';
+import { AllCategory, ErrorInsertInto, Order, Product } from '../types/Product';
 import { promisify } from 'util';
 import { ReqAddToBasket } from "../types/ReqAddToBasket";
 import { OkPacket, RowDataPacketCharacteristic } from "../types/OkPacket";
@@ -129,14 +129,26 @@ class MysqlClient implements SomeDataBase {
             responseCats.categories.push(...rows); // вставим все категории
             // получить все уникальные характеристики по каждой категории
             for (let v of rows) {
-                const charactsQuery = await promCon(`
-                SELECT DISTINCT LCASE(характеристики) AS characteristics FROM Товары WHERE LCASE(характеристики) LIKE "%${v.toLowerCase()}%"; 
-                `) as RowDataPacketCharacteristic[];
                 
-                let strCharacts = charactsQuery[0].characteristics.split('/');
-                const str = strCharacts.slice(0, strCharacts.length-1);
-                responseCats.characteristics[v] = [...str];
+                const charactsQuery = await promCon(`
+                SELECT DISTINCT LCASE(характеристики) AS characteristics FROM Товары WHERE LCASE(характеристики) LIKE "%${v}%"; 
+                `) as RowDataPacketCharacteristic[];
+                let setCharactsQuery = new Set();
+                charactsQuery.forEach(v => {
+                    if (v.characteristics) {
+                        let chr = v.characteristics.split('/');
+                        chr.forEach(el => {
+                            if (el.trim().length > 0)
+                                setCharactsQuery.add(el);
+                        });
+                    }
+                });
+            
+                console.log('set ', setCharactsQuery);
+            
+                responseCats.characteristics[v] = Array.from(setCharactsQuery) as string[];
             } 
+            console.log(responseCats);
             return responseCats;
 
         } catch {
@@ -183,20 +195,35 @@ class MysqlClient implements SomeDataBase {
             if (Array.isArray(result) && result.length > 0) {
                 const p: Product = result[0];
                 const usid = addProd.userId.split('_')[1];
-                const current_basket = await promCon(`SELECT * FROM ${addProd.userId} WHERE product_id=${addProd.idProduct};`) as Product[];
-                
-                if (p['количество_на_складе'] && current_basket.length < +p['количество_на_складе']) {
-                    await promCon(`
-                    INSERT INTO ${addProd.userId} (user_id, product_id, uniq_token, datetime, category, brand,
-                        name_good, characteristic, count_on_stock, price_from_1to2, price_from_3to4, price_from_5to9,
-                        price_from_10to29, price_from_30to69, price_from_70to149, price_from_150, order_status)
-                    VALUES ("${usid}", "${addProd.idProduct}", "", "${new Date().toISOString()}", "${p["группы"].replace(/"/g, '')}", "${p["бренд"].replace(/"/g, '')}", "${p["наименование"].replace(/"/g, '')}",
-                        "${p["характеристики"].replace(/"/g, '')}", "${p["количество_на_складе"] || 0}", "${p["цена_от_1_до_2"] || 0}", "${p["цена_от_3_до_4"] || 0}",
-                        ${p["цена_от_5_до_9"] || 0}, ${p["цена_от_10_до_29"] || 0}, ${p["цена_от_30_до_69"] || 0}, ${p["цена_от_70_до_149"] || 0},
-                        "${p["цена_от_150"] || 0}", "${StatusOrder.IN_BASKET}");
-                    `) as OkPacket;
+                                
+                if ( p['количество_на_складе'] && +p['количество_на_складе'] > 0) {
+                    const findProductWithId = await promCon(`
+                        SELECT count_on_order FROM ${addProd.userId} WHERE product_id=${addProd.idProduct};
+                    `) as Order[];
+
+                    if (Array.isArray(findProductWithId) && findProductWithId.length === 0) {
+                            await promCon(`
+                            INSERT INTO ${addProd.userId} (user_id, product_id, uniq_token, datetime, category, brand,
+                                name_good, characteristic, count_on_stock, count_on_order, price_from_1to2, price_from_3to4, price_from_5to9,
+                                price_from_10to29, price_from_30to69, price_from_70to149, price_from_150, order_status)
+                            VALUES ("${usid}", "${addProd.idProduct}", "", "${new Date().toISOString()}", "${p["группы"].replace(/"/g, '')}", "${p["бренд"].replace(/"/g, '')}", "${p["наименование"].replace(/"/g, '')}",
+                                "${p["характеристики"].replace(/"/g, '')}", "${p["количество_на_складе"] || 0}", "1",  "${p["цена_от_1_до_2"] || 0}", "${p["цена_от_3_до_4"] || 0}",
+                                ${p["цена_от_5_до_9"] || 0}, ${p["цена_от_10_до_29"] || 0}, ${p["цена_от_30_до_69"] || 0}, ${p["цена_от_70_до_149"] || 0},
+                                "${p["цена_от_150"] || 0}", "${StatusOrder.IN_BASKET}");
+                            `) as OkPacket;
+                    }
+                    else {
+                        const findCountOnOrder = +findProductWithId[0].count_on_order;
+                        if (p['количество_на_складе'] && findCountOnOrder < +p['количество_на_складе']) {
+                            await promCon( `
+                                UPDATE ${addProd.userId} 
+                                SET count_on_order="${findCountOnOrder + 1}"
+                                WHERE ${addProd.userId}.product_id=${addProd.idProduct}
+                            ` );
+                        }
                 }
-             
+                }
+
                 const basket = await promCon(`SELECT * FROM ${addProd.userId};`) as Product[];
                 return basket;
             }
@@ -220,8 +247,27 @@ class MysqlClient implements SomeDataBase {
         const promCon = promisify(connect.query).bind(connect);
         try {
             const current_basket = await promCon(`SELECT * FROM ${removeProd.userId} WHERE product_id=${removeProd.idProduct};`) as Product[];
-            if (current_basket.length > 0)
-                await promCon(`DELETE FROM ${removeProd.userId} WHERE product_id=${removeProd.idProduct} LIMIT 1;`) as OkPacket;
+            if (current_basket && current_basket.length > 0) {
+                const findProductWithId = await promCon(`
+                    SELECT count_on_order FROM ${removeProd.userId} WHERE product_id=${removeProd.idProduct};
+                `) as Order[];
+                if (Array.isArray(findProductWithId) && findProductWithId.length > 0) {
+                    const findCountOnOrder = +findProductWithId[0].count_on_order || null;
+                    if (findCountOnOrder) {
+                        if (findCountOnOrder === 1) {
+                            await promCon(`DELETE FROM ${removeProd.userId} WHERE product_id=${removeProd.idProduct} LIMIT 1;`) as OkPacket;
+                        } else {
+                            if (findCountOnOrder > 1)
+                                await promCon( `
+                                        UPDATE ${removeProd.userId} 
+                                        SET count_on_order="${findCountOnOrder - 1}"
+                                        WHERE ${removeProd.userId}.product_id=${removeProd.idProduct}
+                                    ` );
+                        }
+                    }
+                }
+            }
+
         
             const basket = await promCon(`SELECT * FROM ${removeProd.userId};`) as Product[];
             return basket;
